database error logcat

+++ app/src/main/java/com/main/proyek_salez/data/viewmodel/CartViewModel.kt
@@ -33,7 +33,7 @@ class CartViewModel @Inject constructor(
         viewModelScope.launch {
             repository.getCartItems().collect { items ->
                 val cartItemsWithFood = items.mapNotNull { cartItem ->
-                    val foodItem = repository.getFoodItemById(cartItem.foodItemId)
+                    val foodItem = repository.getFoodItemById(cartItem.foodItemId.toInt())
                     foodItem?.let { CartItemWithFood(cartItem, it) }
                 }
                 _cartItems.value = cartItemsWithFood
